requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
asyncpg
pgvector
celery
redis
httpx
slack_sdk
python-dotenv
aiohttp
cryptography
loguru
backoff
openai
numpy